Segmented Content code and refactoring.
Took out more of the 'l_' prefixing name scheme and used sane local variable
names. Also segmeneted the ability to load a Content from a QFile* so that
Content objects aren't restricted to the user's local installation of them.
Also, the XML data representing Contents have changed. It was either that
or me editing all of the pre-packaged XML data.