Spread of dptr to application and removal of the Agent concept in favor of Services.
Removed the use of AbstractAgent in favor of the Services framework, which is
already integrated with the Main window. Also took the opporunity to introduce
the concept of private data for the Desktop Control and Dictation services. The
public API of SpeechControl is beginning to take more of a form, allowing for
more stable bindings for external languages.